Time: 27 ms (67.17%) | Memory: 35.9 MB (68.82%) - LeetSync

+++ 518-coin-change-ii/coin-change-ii.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int change(int amount, vector<int>& coins) {
         int n = coins.size();
-        vector<int> prev(amount + 1, 0);
+        vector<unsigned int> prev(amount + 1, 0);
 
         for (int j = 0; j <= amount; j++) {
             if (j % coins[0] == 0) {
@@ -11,11 +11,11 @@ public:
         }
         for (int i = 1; i < n; i++) {
 
-            vector<int> curr(amount + 1, 0);
+            vector<unsigned int> curr(amount + 1, 0);
             curr[0] = 1;
             for (int j = 1; j <= amount; j++) {
-                size_t notake = prev[j];
-                size_t take = 0;
+                unsigned int notake = prev[j];
+                unsigned int take = 0;
                 if (j >= coins[i]) {
                     take = curr[j - coins[i]];
                 }
